Refactored GlobalButton to GhostButton
