+++ src/components/GhostButton/GhostButton.scss
@@ -1,3 +1,3 @@
-.globalButton {
+.ghostButton {
   border: 1px solid rgb(255, 255, 255);
 }
