+++ src/components/GhostButton/GhostButton.test.js
