+++ src/components/GhostButton/index.js
@@ -1,0 +1,3 @@
+import GhostButton from './GhostButton';
+
+export default GhostButton;
