--- src/components/GlobalButton/index.js
@@ -1,3 +1,0 @@
-import GlobalButton from './GlobalButton';
-
-export default GlobalButton;
